Implment image cleanup

+++ src/Util/File.php
@@ -2,13 +2,24 @@
 
 namespace App\Util;
 
+use App\Exception\CouldNotDeleteFile;
+
 class File
 {
     public function delete(string $filePath) : void
     {
-        if (unlink($filePath) === false) {
-            throw new \RuntimeException('Could not delete file: ' . $filePath);
+        if ($this->fileExist($filePath) === false) {
+            return;
         }
+
+        if (unlink($filePath) === false) {
+            throw new CouldNotDeleteFile('Could not delete file: ' . $filePath);
+        }
+    }
+
+    public function fileExist(string $filePath) : bool
+    {
+        return file_exists($filePath);
     }
 
     public function write(string $filePath, string $content) : void
